pkg/generator: refactor and update test for deploy/* generation
