+++ pkg/generator/generator_test.go
@@ -19,17 +19,6 @@ import (
 	"testing"
 )
 
-const (
-	// test constants for app-operator
-	appRepoPath    = "github.com/example-inc/" + appProjectName
-	appKind        = "AppService"
-	appApiDirName  = "app"
-	appAPIVersion  = appGroupName + "/" + appVersion
-	appVersion     = "v1alpha1"
-	appGroupName   = "app.example.com"
-	appProjectName = "app-operator"
-)
-
 const mainExp = `package main
 
 import (
@@ -377,107 +366,4 @@ func TestGenConfig(t *testing.T) {
 	if configExp != buf.String() {
 		t.Errorf("want %v, got %v", configExp, buf.String())
 	}
-}
-
-const operatorYamlExp = `apiVersion: apiextensions.k8s.io/v1beta1
-kind: CustomResourceDefinition
-metadata:
-  name: appservices.app.example.com
-spec:
-  group: app.example.com
-  names:
-    kind: AppService
-    listKind: AppServiceList
-    plural: appservices
-    singular: appservice
-  scope: Namespaced
-  version: v1alpha1
----
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: app-operator
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      name: app-operator
-  template:
-    metadata:
-      labels:
-        name: app-operator
-    spec:
-      containers:
-        - name: app-operator
-          image: quay.io/coreos/operator-sdk-dev:app-operator
-          command:
-          - app-operator
-          imagePullPolicy: Always
-`
-
-const rbacYamlExp = `kind: Role
-apiVersion: rbac.authorization.k8s.io/v1beta1
-metadata:
-  name: app-operator
-rules:
-- apiGroups:
-  - app.example.com
-  resources:
-  - "*"
-  verbs:
-  - "*"
-- apiGroups:
-  - ""
-  resources:
-  - pods
-  - services
-  - endpoints
-  - persistentvolumeclaims
-  - events
-  - configmaps
-  - secrets
-  verbs:
-  - "*"
-- apiGroups:
-  - apps
-  resources:
-  - deployments
-  - daemonsets
-  - replicasets
-  - statefulsets
-  verbs:
-  - "*"
-
----
-
-kind: RoleBinding
-apiVersion: rbac.authorization.k8s.io/v1beta1
-metadata:
-  name: default-account-app-operator
-subjects:
-- kind: ServiceAccount
-  name: default
-roleRef:
-  kind: Role
-  name: app-operator
-  apiGroup: rbac.authorization.k8s.io
-`
-
-func TestGenDeploy(t *testing.T) {
-	buf := &bytes.Buffer{}
-	if err := renderOperatorYaml(buf, appKind, appAPIVersion, appProjectName, "quay.io/coreos/operator-sdk-dev:app-operator"); err != nil {
-		t.Error(err)
-	}
-	if operatorYamlExp != buf.String() {
-		t.Errorf("want %v, got %v", operatorYamlExp, buf.String())
-	}
-
-	buf = &bytes.Buffer{}
-	if err := renderRBACYaml(buf, appProjectName, appGroupName); err != nil {
-		t.Error(err)
-	}
-	if rbacYamlExp != buf.String() {
-		t.Errorf("want %v, got %v", rbacYamlExp, buf.String())
-	}
-
 }
